+++ apiserver/paasng/paas_wl/bk_app/monitoring/metrics/models.py
@@ -23,8 +23,7 @@ from typing import TYPE_CHECKING, Generator, List, Optional
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
 
-from paas_wl.infras.cluster.constants import ClusterFeatureFlag
-from paas_wl.infras.cluster.utils import get_cluster_by_app
+from paas_wl.bk_app.applications.models import WlApp
 from paas_wl.bk_app.monitoring.metrics.clients import (
     BkMonitorMetricClient,
     MetricClient,
@@ -35,11 +34,12 @@ from paas_wl.bk_app.monitoring.metrics.clients import (
 from paas_wl.bk_app.monitoring.metrics.constants import MetricsResourceType, MetricsSeriesType
 from paas_wl.bk_app.monitoring.metrics.exceptions import AppInstancesNotFoundError, AppMetricNotSupportedError
 from paas_wl.bk_app.monitoring.metrics.utils import MetricSmartTimeRange
-from paas_wl.bk_app.applications.models import WlApp
 from paas_wl.bk_app.processes.readers import instance_kmodel, process_kmodel
+from paas_wl.infras.cluster.constants import ClusterFeatureFlag
+from paas_wl.infras.cluster.utils import get_cluster_by_app
 
 if TYPE_CHECKING:
-    from paas_wl.bk_app.processes.entities import Process
+    from paas_wl.bk_app.processes.kres_entities import Process
 
 logger = logging.getLogger(__name__)
