+++ apiserver/paasng/paas_wl/bk_app/processes/controllers.py
@@ -25,8 +25,8 @@ from django.utils.translation import ugettext_lazy as _
 
 from paas_wl.bk_app.applications.constants import WlAppType
 from paas_wl.bk_app.applications.models import Release, WlApp
-from paas_wl.bk_app.processes.entities import Process
 from paas_wl.bk_app.processes.exceptions import ProcessOperationTooOften
+from paas_wl.bk_app.processes.kres_entities import Process
 from paas_wl.bk_app.processes.models import ProcessSpec
 from paas_wl.bk_app.processes.readers import instance_kmodel, ns_instance_kmodel, ns_process_kmodel, process_kmodel
 from paas_wl.workloads.autoscaling.entities import AutoscalingConfig
