+++ apiserver/paasng/paas_wl/bk_app/processes/readers.py
@@ -21,6 +21,8 @@ from typing import Dict, List, Optional, cast
 from django.core.exceptions import ObjectDoesNotExist
 from kubernetes.dynamic import ResourceInstance
 
+from paas_wl.bk_app.applications.constants import WlAppType
+from paas_wl.bk_app.applications.models import WlApp
 from paas_wl.bk_app.cnative.specs.constants import (
     BKAPP_CODE_ANNO_KEY,
     ENVIRONMENT_ANNO_KEY,
@@ -28,14 +30,12 @@ from paas_wl.bk_app.cnative.specs.constants import (
     RESOURCE_TYPE_KEY,
     WLAPP_NAME_ANNO_KEY,
 )
+from paas_wl.bk_app.processes.kres_entities import Instance, Process
 from paas_wl.core.resource import get_process_selector
-from paas_wl.bk_app.applications.constants import WlAppType
-from paas_wl.bk_app.applications.models import WlApp
 from paas_wl.infras.resources.base.exceptions import NotAppScopedResource
 from paas_wl.infras.resources.base.kres import KPod
 from paas_wl.infras.resources.kube_res.base import AppEntityReader, NamespaceScopedReader, ResourceList
 from paas_wl.infras.resources.kube_res.exceptions import AppEntityNotFound
-from paas_wl.bk_app.processes.entities import Instance, Process
 from paasng.platform.applications.models import ModuleEnvironment
