+++ apiserver/paasng/paas_wl/bk_app/processes/serializers.py
@@ -16,338 +16,383 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
-#########################
-# Serializers & Readers #
-#########################
-from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple
+import copy
+from dataclasses import dataclass, field
+from typing import Any, Dict, Generic, List, Optional, TypeVar
+from uuid import UUID
 
+import arrow
 import cattr
-from django.conf import settings
-from kubernetes.dynamic import ResourceField, ResourceInstance
+from django.utils.translation import ugettext_lazy as _
+from kubernetes.utils.quantity import parse_quantity
+from rest_framework import serializers
+from rest_framework.serializers import ValidationError
 
-from paas_wl.bk_app.applications.constants import WlAppType
-from paas_wl.bk_app.applications.models import Release, WlApp
-from paas_wl.bk_app.processes.constants import PROCESS_MAPPER_VERSION_KEY, PROCESS_NAME_KEY
-from paas_wl.bk_app.processes.exceptions import UnknownProcessTypeError
-from paas_wl.infras.cluster.utils import get_cluster_by_app
-from paas_wl.infras.resource_templates.logging import get_app_logging_volume, get_app_logging_volume_mounts
-from paas_wl.infras.resource_templates.utils import AddonManager, ProcessProbeManager
-from paas_wl.infras.resources.kube_res.base import AppEntityDeserializer, AppEntitySerializer
-from paas_wl.utils.kubestatus import HealthStatus, HealthStatusType, check_pod_health_status, parse_pod
-from paas_wl.workloads.release_controller.constants import ImagePullPolicy
-
-if TYPE_CHECKING:
-    from paas_wl.bk_app.processes.entities import Instance, Process
-    from paas_wl.infras.resources.generation.mapper import MapperPack
-
-
-def extract_type_from_name(name: str, namespace: str) -> Optional[str]:
-    """Extract process type from Pod/Deployment name. This function exists for compatibility reasons.
-
-    While each deployment has a `process_id` label defined in `pod.metadata.labels`, in older version, the
-    label does not exists, so the only way is extracting process_type from resource name.
-    """
-    try:
-        return name.split(namespace)[-1].split("-")[1]
-    except IndexError:
-        return None
+from paas_wl.bk_app.applications.models import Release
+from paas_wl.bk_app.processes.constants import ProcessUpdateType
+from paas_wl.bk_app.processes.kres_entities import Instance, Process
+from paas_wl.bk_app.processes.models import ProcessSpec
+from paas_wl.infras.resources.kube_res.base import WatchEvent
+from paas_wl.workloads.autoscaling.constants import ScalingMetric, ScalingMetricSourceType
+from paas_wl.workloads.autoscaling.entities import AutoscalingConfig
+from paas_wl.workloads.networking.ingress.utils import get_service_dns_name
+from paasng.platform.engine.constants import JobStatus, RuntimeType
+from paasng.platform.engine.models import Deployment, OfflineOperation
+from paasng.platform.sourcectl.models import VersionInfo
 
 
-class ProcessDeserializer(AppEntityDeserializer['Process']):
-    """Deserializer for Process"""
+class HumanizeDateTimeField(serializers.DateTimeField):
+    def to_representation(self, value):
+        return arrow.get(value).humanize(locale="zh")
 
-    def deserialize(self, app: WlApp, kube_data: ResourceInstance) -> 'Process':
-        """Get meta info from pod template
-        :param app: workloads app
-        :param kube_data: k8s Deployment
-        :return: Process
+
+class ProcessForDisplaySLZ(serializers.Serializer):
+    """Common serializer for representing Process object"""
+
+    # Warning: 这个字段需要查询数据库
+    module_name = serializers.CharField(source="app.module_name", help_text="模块名")
+
+    name = serializers.CharField(source='metadata.name')
+    type = serializers.CharField(help_text="进程类型(名称)")
+    command = serializers.CharField(source="runtime.proc_command", help_text="进程命令")
+
+    replicas = serializers.IntegerField(source="status.replicas", help_text="期望实例数")
+    success = serializers.IntegerField(source="status.success", help_text="成功进程个数")
+    failed = serializers.IntegerField(source="status.failed", help_text="失败进程个数")
+    version = serializers.IntegerField(help_text="release version")
+
+    cluster_link = serializers.SerializerMethodField()
+
+    def get_cluster_link(self, process: Process) -> str:
+        return f'http://{get_service_dns_name(process.app, process.type)}'
+
+
+class InstanceForDisplaySLZ(serializers.Serializer):
+    """Common serializer for representing Instance object, removes some extra
+    large and sensitive fields such as "envs" """
+
+    # Warning: 这个字段需要查询数据库
+    module_name = serializers.CharField(source="app.module_name")
+
+    name = serializers.CharField(read_only=True)
+    process_type = serializers.CharField(read_only=True)
+    display_name = serializers.CharField(read_only=True, source="shorter_name")
+    image = serializers.CharField(read_only=True)
+    start_time = serializers.CharField(read_only=True)
+    state = serializers.SerializerMethodField(read_only=True, help_text='实例状态')
+    state_message = serializers.CharField(read_only=True)
+    ready = serializers.BooleanField(read_only=True)
+    restart_count = serializers.IntegerField()
+    version = serializers.CharField(read_only=True)
+
+    def get_state(self, obj: Instance) -> str:
+        """Sometimes the target pod may enter into 'Running' phase without being ready,
+        which may confuses users, translate those states into "Starting"
         """
-        if app.type == WlAppType.DEFAULT:
-            process = self._deserialize_for_default_app(app, kube_data)
-        else:
-            process = self._deserialize_for_cnative_app(app, kube_data)
+        if not obj.ready and obj.state == "Running":
+            # Not using word "Pending" because it's already an Kubernetes Pod state
+            return "Starting"
+        return obj.state
 
-        # failed should never be negative
-        process.fulfill_runtime(
-            replicas=kube_data.spec.replicas,
-            success=kube_data.status.get('availableReplicas') or 0,
-            metadata=kube_data.metadata,
-        )
 
-        return process
+class ProcessExtraInfoSLZ(serializers.Serializer):
+    """part of SLZ for ProcessInstanceListWatcher.list"""
 
-    def _deserialize_for_default_app(self, app: WlApp, kube_data: ResourceInstance) -> 'Process':
-        """deserialize process info for default type(Heroku) app"""
-        from paas_wl.bk_app.processes.entities import Resources, Runtime, Schedule
+    type = serializers.CharField(help_text="进程类型")
+    command = serializers.CharField(help_text="进程命令")
+    cluster_link = serializers.CharField(help_text="集群内访问地址")
 
-        main_container = self._get_main_container(app, kube_data)
-        pod_spec = kube_data.spec.template.spec
-        pod_labels = kube_data.spec.template.metadata.labels
-        version = int(pod_labels.get('release_version', 0))
-        process_type = self._get_process_type(kube_data)
+
+class ListRespMetaDataSLZ(serializers.Serializer):
+    """part of SLZ for ProcessInstanceListWatcher.list"""
+
+    resource_version = serializers.CharField(help_text="k8s 资源版本")
+
+
+class ProcessListSLZ(serializers.Serializer):
+    """part of SLZ for ProcessInstanceListWatcher.list"""
+
+    items = ProcessForDisplaySLZ(many=True)
+    extra_infos = ProcessExtraInfoSLZ(many=True, required=False, help_text="[deprecated] 相关信息移动到 ProcessForDisplaySLZ")
+    metadata = ListRespMetaDataSLZ()
+
+
+class InstanceListSLZ(serializers.Serializer):
+    """part of SLZ for ProcessInstanceListWatcher.list"""
+
+    items = InstanceForDisplaySLZ(many=True)
+    metadata = ListRespMetaDataSLZ()
+
+
+class ListWatcherRespSLZ(serializers.Serializer):
+    """SLZ for ProcessInstanceListWatcher.list"""
+
+    processes = ProcessListSLZ()
+    instances = InstanceListSLZ()
+    cnative_proc_specs = serializers.ListField(required=False, child=serializers.DictField())
+    process_packages = serializers.ListField(required=False, child=serializers.DictField())
+
+
+class VersionInfoSLZ(serializers.Serializer):
+    """已部署的代码/镜像版本信息"""
+
+    revision = serializers.CharField(help_text="版本hash")
+    version_name = serializers.CharField(help_text="版本名称")
+    version_type = serializers.CharField(help_text="版本类型(tag/branch)")
+
+
+class DeploymentOperationSLZ(serializers.Serializer):
+    """部署操作对象"""
+
+    id = serializers.CharField(help_text="部署ID")
+    status = serializers.ChoiceField(choices=JobStatus.get_choices())
+    created = serializers.DateTimeField(help_text="操作时间")
+
+    version_info = VersionInfoSLZ(help_text="版本信息")
+    err_detail = serializers.CharField(help_text="错误原因")
+    has_requested_int = serializers.BooleanField(help_text="部署是否被中断")
+
+
+class OfflineOperationSLZ(serializers.Serializer):
+    """下架操作对象"""
+
+    id = serializers.CharField(help_text="下架ID")
+    status = serializers.ChoiceField(choices=JobStatus.get_choices())
+    created = serializers.DateTimeField(help_text="操作时间")
+
+
+def make_operation_group_slz(child: serializers.Serializer):
+    return type(
+        f"{type(child).__name__}Group",
+        (serializers.Serializer,),
+        {
+            "latest": copy.deepcopy(child),
+            "latest_succeeded": copy.deepcopy(child),
+            "pending": copy.deepcopy(child),
+        },
+    )()
+
+
+class ModuleStateSLZ(serializers.Serializer):
+    """和模块相关的部署/下架对象"""
+
+    deployment = make_operation_group_slz(DeploymentOperationSLZ(help_text="部署操作对象", allow_null=True))
+    offline = make_operation_group_slz(OfflineOperationSLZ(help_text="下架操作对象", allow_null=True))
+
+
+class NamespaceScopedListWatchRespPartSLZ(serializers.Serializer):
+    module_name = serializers.CharField()
+    state = ModuleStateSLZ(help_text="模块状态")
+    exposed_url = serializers.CharField(required=False, help_text="访问地址")
+    repo_url = serializers.CharField(required=False, help_text="源码/镜像仓库地址")
+    build_method = serializers.CharField(help_text="构建方式")
+    version_info = VersionInfoSLZ(required=False)
+
+    processes = ProcessForDisplaySLZ(many=True)
+    instances = InstanceForDisplaySLZ(many=True)
+    proc_specs = serializers.ListField(child=serializers.DictField())
+
+    total_available_instance_count = serializers.IntegerField(help_text="总运行实例数")
+    total_desired_replicas = serializers.IntegerField(help_text="总期望实例数")
+    total_failed = serializers.IntegerField(help_text="总异常实例数")
+
+
+OP = TypeVar("OP")
+
+
+@dataclass
+class OperationGroup(Generic[OP]):
+    latest: Optional[OP] = None
+    latest_succeeded: Optional[OP] = None
+    pending: Optional[OP] = None
+
+
+@dataclass
+class ModuleState:
+    deployment: OperationGroup[Deployment]
+    offline: OperationGroup[OfflineOperation]
+
+
+@dataclass
+class ModuleScopedData:
+    module_name: str
+    build_method: RuntimeType
+    state: ModuleState
+    exposed_url: Optional[str]
+    repo_url: Optional[str]
+    version_info: Optional[VersionInfo] = None
+
+    processes: List[Process] = field(default_factory=list)
+    instances: List[Instance] = field(default_factory=list)
+    proc_specs: List[Dict] = field(default_factory=list)
+
+    @property
+    def total_available_instance_count(self) -> int:
+        return sum(p.status.success for p in self.processes if p.status)
+
+    @property
+    def total_desired_replicas(self) -> int:
+        # Q: 为什么不使用 process_specs 的 target_replicas
+        # A: 因为该字段对自动扩缩容的进程无效
+        return sum(p.status.replicas for p in self.processes if p.status)
+
+    @property
+    def total_failed(self) -> int:
+        return sum(p.status.failed for p in self.processes if p.status)
+
+
+class NamespaceScopedListWatchRespSLZ(serializers.Serializer):
+    data = NamespaceScopedListWatchRespPartSLZ(many=True)
+    rv_proc = serializers.CharField(required=True)
+    rv_inst = serializers.CharField(required=True)
+
+
+class ErrorEventSLZ(serializers.Serializer):
+    """SLZ for WatchEvent which type == 'ERROR'"""
+
+    type = serializers.CharField()
+    error_message = serializers.CharField()
+
+
+class WatchEventSLZ(serializers.Serializer):
+    """SLZ for ProcessInstanceListWatcher.watch"""
+
+    type = serializers.CharField()
+    object_type = serializers.CharField()
+    object = serializers.DictField()
+    resource_version = serializers.CharField(allow_null=True, help_text="仅 object_type != 'error' 时有该字段")
+
+    def to_representation(self, instance: WatchEvent):
+        data: Dict[str, Any] = {"type": instance.type}
+        if instance.type == 'ERROR':
+            data["object_type"] = "error"
+            data["object"] = ErrorEventSLZ(instance).data
+        if isinstance(instance.res_object, Process):
+            data["object_type"] = "process"
+            data["object"] = ProcessForDisplaySLZ(instance.res_object).data
+            data["resource_version"] = instance.res_object.get_resource_version()
+        elif isinstance(instance.res_object, Instance):
+            data["object_type"] = "instance"
+            data["object"] = InstanceForDisplaySLZ(instance.res_object).data
+            data["resource_version"] = instance.res_object.get_resource_version()
+        return super().to_representation(data)
+
+
+class ProcessSpecSLZ(serializers.Serializer):
+    """Serializer for representing process packages
+
+    Need to convert the resource limit to a number:
+    "resource_limit": {
+        "cpu": "100m",
+        "memory": "64Mi"
+    }
+    "resource_limit_quota": {
+        "cpu": 100,
+        "memory": 64
+    }
+    """
+
+    name = serializers.CharField()
+    target_replicas = serializers.IntegerField()
+    target_status = serializers.CharField()
+    max_replicas = serializers.IntegerField(source='plan.max_replicas')
+    resource_limit = serializers.JSONField(source='plan.limits')
+    resource_requests = serializers.JSONField(source='plan.requests')
+    plan_name = serializers.CharField(source="plan.name")
+    resource_limit_quota = serializers.SerializerMethodField(read_only=True)
+    autoscaling = serializers.BooleanField()
+    scaling_config = serializers.JSONField()
+
+    # deprecated: 兼容旧的 CNativeProcSpecSLZ, 待前端替换取值字段后移除
+    cpu_limit = serializers.CharField(source="plan.limits.cpu")
+    memory_limit = serializers.CharField(source="plan.limits.memory")
+
+    def get_resource_limit_quota(self, obj: ProcessSpec) -> dict:
+        limits = obj.plan.limits
+        # 内存的单位为 Mi
+        memory_quota = int(parse_quantity(limits['memory']) / (1024 * 1024))
+        # CPU 的单位为 m
+        cpu_quota = int(parse_quantity(limits['cpu']) * 1000)
+        return {"cpu": cpu_quota, "memory": memory_quota}
+
+
+class ScalingObjectRefSLZ(serializers.Serializer):
+    """资源引用"""
+
+    api_version = serializers.CharField(required=True)
+    kind = serializers.CharField(required=True)
+    name = serializers.CharField(required=True)
+
+
+class MetricSpecSLZ(serializers.Serializer):
+    """扩缩容指标"""
+
+    type = serializers.ChoiceField(
+        required=False,
+        default=ScalingMetricSourceType.RESOURCE,
+        choices=ScalingMetricSourceType.get_choices(),
+    )
+    metric = serializers.ChoiceField(required=True, choices=ScalingMetric.get_choices())
+    value = serializers.CharField(required=True, help_text=_('资源指标值/百分比'))
+    described_object = ScalingObjectRefSLZ(required=False)
+
+
+class ScalingConfigSLZ(serializers.Serializer):
+    """扩缩容配置"""
+
+    min_replicas = serializers.IntegerField(required=True, min_value=1, help_text=_('最小副本数'))
+    max_replicas = serializers.IntegerField(required=True, min_value=1, help_text=_('最大副本数'))
+    metrics = serializers.ListField(child=MetricSpecSLZ(), required=True, min_length=1, help_text=_('扩缩容指标'))
+
+
+class UpdateProcessSLZ(serializers.Serializer):
+    """Serializer for updating processes"""
+
+    process_type = serializers.CharField(required=True)
+    operate_type = serializers.ChoiceField(required=True, choices=ProcessUpdateType.get_django_choices())
+    target_replicas = serializers.IntegerField(required=False, min_value=1, help_text=_('目标进程副本数'))
+    autoscaling = serializers.BooleanField(required=False, default=False, help_text=_('是否开启自动扩缩容'))
+    scaling_config = ScalingConfigSLZ(required=False, help_text=_('进程扩缩容配置'))
+
+    def validate(self, attrs: Dict) -> Dict:
+        if attrs['operate_type'] == ProcessUpdateType.SCALE:
+            if attrs['autoscaling']:
+                if not attrs.get('scaling_config'):
+                    raise ValidationError(_('当启用自动扩缩容时，必须提供有效的 scaling_config'))
+            elif not attrs.get('target_replicas'):
+                raise ValidationError(_('当操作类型为扩缩容时，必须提供有效的 target_replicas'))
+
+        return attrs
+
+    def validate_scaling_config(self, config: Dict) -> Optional[AutoscalingConfig]:
+        if not config:
+            return None
+        try:
+            return cattr.structure(config, AutoscalingConfig)
+        except Exception as e:
+            raise ValidationError(_('scaling_config 配置格式有误：{}').format(e))
+
+
+class ListProcessesQuerySLZ(serializers.Serializer):
+    """Serializer for query params of list API"""
+
+    release_id = serializers.UUIDField(default=None, help_text="用于过滤实例的发布ID")
+
+    def validate_release_id(self, release_id: Optional[UUID]) -> Optional[UUID]:
+        """Validate release_id by querying database"""
+        if not release_id:
+            return None
 
         try:
-            release = Release.objects.get_by_version(app, version=version)
-            proc_command = release.get_procfile()[process_type]
-        except (KeyError, Release.DoesNotExist):
-            proc_command = ""
-
-        process = self.entity_type(
-            app=app,
-            type=process_type,
-            name=kube_data.metadata.name,
-            version=version,
-            replicas=kube_data.spec.replicas,
-            runtime=cattr.structure(
-                {
-                    "envs": {env.name: env.value for env in main_container.env if getattr(env, "value", None)},
-                    "image": main_container.image,
-                    "command": getattr(main_container, "command", []),
-                    "args": getattr(main_container, "args", []),
-                    "proc_command": proc_command,
-                    "image_pull_policy": getattr(main_container, "imagePullPolicy", ImagePullPolicy.IF_NOT_PRESENT),
-                    "image_pull_secrets": getattr(pod_spec, "imagePullSecrets", []),
-                },
-                Runtime,
-            ),
-            schedule=cattr.structure(
-                {
-                    "cluster_name": get_cluster_by_app(app).name,
-                    "tolerations": getattr(pod_spec, "tolerations", []),
-                    "node_selector": getattr(pod_spec, "nodeSelector", {}),
-                },
-                Schedule,
-            ),
-            resources=cattr.structure(getattr(main_container, "resources", None), Resources),
-        )
-        return process
-
-    def _deserialize_for_cnative_app(self, app: WlApp, kube_data: ResourceInstance) -> 'Process':
-        """deserialize process info for cloud native type app"""
-        from paas_wl.bk_app.processes.entities import Runtime, Schedule
-
-        main_container = self._get_main_container(app, kube_data)
-        process = self.entity_type(
-            app=app,
-            name=self._get_process_type(kube_data),
-            version=0,
-            replicas=kube_data.spec.replicas,
-            type=self._get_process_type(kube_data),
-            schedule=Schedule(
-                cluster_name=app.latest_config.cluster,
-                # TODO: 从 pod template 解析这些属性?
-                tolerations=[],
-                node_selector={},
-            ),
-            runtime=Runtime(
-                # TODO: 从 pod template 解析 env
-                envs={},
-                image=main_container.image,
-                command=getattr(main_container, "command", []),
-                args=getattr(main_container, "args", []),
-                image_pull_policy=main_container.imagePullPolicy,
-            ),
-        )
-        return process
-
-    @staticmethod
-    def _get_process_type(deployment: ResourceInstance) -> str:
-        """Get process type for deployment resource
-
-        :raises: UnknownProcessTypeError: when no process_type info can be found
-        """
-        process_type = deployment.spec.template.metadata.labels.get(PROCESS_NAME_KEY)
-        if process_type:
-            return process_type
-
-        # label `process_id` is deprecated, should use `PROCESS_NAME_KEY` instead
-        process_type = deployment.spec.template.metadata.labels.get('process_id')
-        if process_type:
-            return process_type
-
-        process_type = extract_type_from_name(deployment.metadata.name, deployment.metadata.namespace)
-        if process_type:
-            return process_type
-
-        raise UnknownProcessTypeError(res=deployment, msg="'No process_type found in resource")
-
-    def _get_main_container(self, app: WlApp, deployment: ResourceInstance) -> ResourceField:
-        """Get main container from main Pod"""
-        process_type = self._get_process_type(deployment)
-        pod_template = deployment.spec.template
-        for c in pod_template.spec.containers:
-            if c.name == app.scheduler_safe_name_with_region:
-                return c
-
-            if c.name == process_type:
-                return c
-        raise RuntimeError("No main container found in resource")
+            self.context['wl_app'].release_set.get(pk=release_id)
+        except Release.DoesNotExist:
+            raise ValidationError(f'Release with id={release_id} do not exists')
+        return release_id
 
 
-class InstanceDeserializer(AppEntityDeserializer['Instance']):
-    """Deserializer for Instance"""
+class WatchProcessesQuerySLZ(serializers.Serializer):
+    """Serializer for query params of watch API"""
 
-    def deserialize(self, app: WlApp, kube_data: ResourceInstance) -> 'Instance':
-        """Generate a ProcInstance by given Pod object"""
-        pod = kube_data
-        health_status = check_pod_health_status(parse_pod(kube_data))
-        instance_state, state_message = self.parse_instance_state(pod.status.phase, health_status)
-
-        # Use first container's status
-        c_status = None
-        if pod.status.get('containerStatuses'):
-            c_status = pod.status.containerStatuses[0]
-
-        process_type = self.get_process_type(pod)
-        target_container = self._get_main_container(app, pod)
-
-        envs = {}
-        if target_container and hasattr(target_container, "env"):
-            for env in target_container.env:
-                name = getattr(env, "name", None)
-                value = getattr(env, "value", None)
-                if name and value is not None:
-                    envs[name] = value
-
-        return self.entity_type(
-            app=app,
-            name=pod.metadata.name,
-            process_type=process_type,
-            host_ip=pod.status.get('hostIP', None),
-            start_time=pod.status.get('startTime', None),
-            state=instance_state,
-            state_message=state_message,
-            image=target_container.image if target_container else "",
-            envs=envs,
-            ready=health_status.status == HealthStatusType.HEALTHY,
-            restart_count=c_status.restartCount if c_status else 0,
-            version=int(pod.metadata.labels.get('release_version', 0)),
-        )
-
-    def _get_main_container(self, app: WlApp, pod_info: ResourceInstance) -> Optional[ResourceField]:
-        process_type = self.get_process_type(pod_info)
-        for c in pod_info.spec.containers:
-            if c.name == app.scheduler_safe_name_with_region:
-                return c
-
-            if c.name == process_type:
-                return c
-        return None
-
-    @staticmethod
-    def get_process_type(pod: ResourceInstance) -> str:
-        """Get process type for pod resource
-
-        :raises: UnknownProcessTypeError: when no process_type info can be found
-        """
-        labels = getattr(pod.metadata, "labels", {})
-        process_type = labels.get(PROCESS_NAME_KEY)
-        if process_type:
-            return process_type
-
-        # label `process_id` is deprecated, should use `PROCESS_NAME_KEY` instead
-        process_type = labels.get('process_id')
-        if process_type:
-            return process_type
-
-        process_type = extract_type_from_name(pod.metadata.name, pod.metadata.namespace)
-        if process_type:
-            return process_type
-
-        raise UnknownProcessTypeError(res=pod, msg="No process_type found in resource")
-
-    @staticmethod
-    def parse_instance_state(pod_phase: str, health_status: HealthStatus) -> Tuple[str, str]:
-        if health_status.status == HealthStatusType.UNHEALTHY:
-            return health_status.reason or "Failed", health_status.message
-        elif health_status.status == HealthStatusType.PROGRESSING:
-            return "Pending", health_status.message
-        elif health_status.status == HealthStatusType.HEALTHY:
-            if pod_phase == "Succeeded":
-                return "Succeeded", health_status.message
-            return "Running", health_status.message
-        return health_status.reason or "Unknown", health_status.message
-
-
-class ProcessSerializer(AppEntitySerializer['Process']):
-    """Serializer for process"""
-
-    def serialize(self, obj: 'Process', original_obj: Optional[ResourceInstance] = None, **kwargs) -> Dict:
-        mapper_version: Optional['MapperPack'] = kwargs.get("mapper_version")
-        if mapper_version is None:
-            raise ValueError("mapper_version is required")
-
-        deployment_body: Dict[str, Any] = {
-            'metadata': {
-                'labels': mapper_version.deployment(process=obj).labels,
-                'name': obj.name,
-                'annotations': {PROCESS_MAPPER_VERSION_KEY: mapper_version.version},
-            },
-            'spec': {
-                'revisionHistoryLimit': settings.MAX_RS_RETAIN,
-                # add rolling update strategy to avoid 502 when redeploy or rolling update app
-                'strategy': {
-                    'type': 'RollingUpdate',
-                    'rollingUpdate': {
-                        'maxUnavailable': 0,
-                        # speed up for those processes which own multiple replicas
-                        'maxSurge': '75%',
-                    },
-                },
-                'selector': {'matchLabels': mapper_version.deployment(process=obj).match_labels},
-                'minReadySeconds': 1,
-                'template': {
-                    'spec': self._construct_pod_body_specs(obj),
-                    'metadata': {
-                        'labels': mapper_version.pod(process=obj).labels,
-                        'name': mapper_version.pod(process=obj).name,
-                    },
-                },
-                'replicas': obj.replicas,
-            },
-            'apiVersion': self.get_apiversion(),
-            'kind': 'Deployment',
-        }
-        return deployment_body
-
-    def _construct_pod_body_specs(self, process: 'Process') -> Dict:
-        addon_mgr = AddonManager(process.app)
-        process_probe_mgr = ProcessProbeManager(app=process.app, process_type=process.type)
-        readiness_probe = cattr.unstructure(process_probe_mgr.get_readiness_probe())
-        if readiness_probe is None and process.type == "web":
-            readiness_probe = cattr.unstructure(addon_mgr.get_readiness_probe())
-        liveness_probe = cattr.unstructure(process_probe_mgr.get_liveness_probe())
-        startup_probe = cattr.unstructure(process_probe_mgr.get_startup_probe())
-        main_container = {
-            'env': [{"name": str(key), "value": str(value)} for key, value in process.runtime.envs.items()],
-            # add preStop to avoid 502 when redeploy or rolling update app
-            'lifecycle': {'pre_stop': {'_exec': {'command': ["sleep", "15"]}}},
-            'image': process.runtime.image,
-            # TODO: 与 cnative 应用统一主容器名字
-            'name': process.main_container_name,
-            'command': process.runtime.command,
-            'args': process.runtime.args,
-            'imagePullPolicy': process.runtime.image_pull_policy,
-            'resources': {
-                'limits': process.resources.limits if process.resources else {},
-                'requests': process.resources.requests if process.resources else {},
-            },
-            # TODO: 重构「主入口」时, 允许用户自行填写 containerPort
-            'ports': [{'containerPort': settings.CONTAINER_PORT}],
-            'volumeMounts': cattr.unstructure(
-                get_app_logging_volume_mounts(process.app) + addon_mgr.get_volume_mounts()
-            ),
-            'readinessProbe': readiness_probe,
-            'livenessProbe': liveness_probe,
-            'startupProbe': startup_probe,
-        }
-
-        return {
-            'containers': [main_container] + cattr.unstructure(addon_mgr.get_sidecars()),
-            'volumes': cattr.unstructure(get_app_logging_volume(process.app) + addon_mgr.get_volumes()),
-            'imagePullSecrets': process.runtime.image_pull_secrets,
-            'nodeSelector': process.schedule.node_selector or None,
-            'tolerations': process.schedule.tolerations or None,
-            # 不默认向 Pod 中挂载 ServiceAccount Token
-            'automountServiceAccountToken': False,
-        }
+    timeout_seconds = serializers.IntegerField(required=False, default=30, max_value=120)
+    rv_proc = serializers.CharField(required=True)
+    rv_inst = serializers.CharField(required=True)
