+++ apiserver/paasng/paas_wl/bk_app/processes/shim.py
@@ -22,10 +22,10 @@ from django.utils.functional import cached_property
 
 from paas_wl.bk_app.applications.models import WlApp
 from paas_wl.bk_app.processes.controllers import Process, list_processes
-from paas_wl.bk_app.processes.drf_serializers import ProcessSpecSLZ
 from paas_wl.bk_app.processes.models import ProcessSpecManager, ProcessTmpl
 from paas_wl.bk_app.processes.processes import PlainProcess, condense_processes
 from paas_wl.bk_app.processes.readers import process_kmodel
+from paas_wl.bk_app.processes.serializers import ProcessSpecSLZ
 from paas_wl.infras.cluster.utils import get_cluster_by_app
 from paas_wl.infras.resources.base.bcs_client import BCSClient
 from paasng.platform.applications.models import ModuleEnvironment
